Added first draft of the project
This was my first attempt at text scraping, and in coming iterations I'll work to make this code more generalized, abstract, and replicable.

This draft was built initially outside of github, so this first commit includees a huge chunk of code. In subsequent versions, my committs will clean this code up and make it more legible and flexible.